% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean object names}
\source{
Slightly modified copy of \code{old_make_clean_names()} from the
\href{https://github.com/sfirke/janitor}{{janitor}} package (MIT license).
}
\usage{
clean_names(dat)
}
\arguments{
\item{dat}{Input object, typically a \code{data.frame}.}
}
\value{
Returns \code{dat} with new, clean names.
}
\description{
Resulting names of object are unique and consist only of
underscores, numbers and letters.
}
\examples{
df <- data.frame(
  `This.is__a...\%..Really_horrific_name` = 1:10,
  `and....this is also (really) bad_` = 1:10
) |> 
  clean_names()
  
names(df)

}
